requirements.txt
# Satellite Imagery Export Tool - Requirements
# Self-Contained Edition - Works on Windows, Linux, macOS, and NixOS!

# GUI Framework
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0

# Geospatial library (has pre-built wheels - installs easily!)
rasterio>=1.3.0

# Image processing
Pillow>=8.0.0

# Numerical computing
numpy>=1.20.0

# HTTP requests for tile downloading
requests>=2.25.0

# Installation Instructions:
# ========================
#
# Simple one-command install on ALL platforms:
#
# Windows / Linux / macOS / NixOS:
#   pip install -r requirements.txt
#
# That's it! No complex setup needed.
#
# Why rasterio instead of GDAL?
# - Rasterio has pre-built binary wheels for all platforms
# - Includes GDAL functionality we need
# - Installs via pip with zero configuration
# - Works out of the box on Windows, Linux, macOS, and NixOS
#
# If pip install fails (rare), use conda:
#   conda install -c conda-forge rasterio pyqt pillow requests numpy
